Started moving server code into spaceapi_server crate
