--- spaceapi-server/src/lib.rs
@@ -1,3 +1,0 @@
-#[test]
-fn it_works() {
-}
