+++ spaceapi/src/lib.rs
@@ -128,6 +128,41 @@ pub struct Status {
 
 }
 
+impl Status {
+
+    /// Create a new Status object with only the minimum amount of fields
+    pub fn new(space: String, logo: String, url: String, location: Location, contact: Contact, issue_report_channels: Vec<String>) -> Status {
+        Status {
+            api: "0.13".to_string(),
+            space: space,
+            logo: logo,
+            url: url,
+            location: location,
+            contact: contact,
+
+            spacefed: Optional::Absent,
+            projects: Optional::Absent,
+            cam: Optional::Absent,
+            feeds: Optional::Absent,
+            events: Optional::Absent,
+            radio_show: Optional::Absent,
+
+            cache: Optional::Absent,
+            issue_report_channels: issue_report_channels,
+
+            state: State {
+                open: None,
+                lastchange: Optional::Absent,
+                trigger_person: Optional::Absent,
+                message: Optional::Absent,
+                icon: Optional::Absent,
+            },
+            sensors: Optional::Absent,
+        }
+    }
+
+}
+
 impl ToJson for Status {
     fn to_json(&self) -> Json {
         let mut d = BTreeMap::new();
