+++ spaceapi_server/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "spaceapi-server"
+name = "spaceapi_server"
 version = "0.1.0"
 authors = [
     "Raphael Nestler <raphael.nestler@gmail.com>",
@@ -9,6 +9,7 @@ authors = [
 [dependencies]
 redis = "0.3"
 hyper = "0.3"
+rustc-serialize = "0.3"
 
 [dependencies.spaceapi]
 path = "../spaceapi"
